+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # Changelog
 
 ## [unreleased]
+### Features
+ - Download script and respective instructions to download .fit files to desired folder